+++ LibrarySystem/LibrarySystem/Program.cs
@@ -1,5 +1,4 @@
 using LibrarySystem.Data;
-using LibrarySystem.Infrastructure;
 using LibrarySystem.Infrastructure.Interfaces;
 using LibrarySystem.Infrastructure.Services;
 using LibrarySystem.Models;
